requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.4
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.2.0
